gemini-microservice/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
pydantic==2.10.3
python-decouple==3.8
python-dotenv==1.0.1
google-genai==1.41.0
httpx
pillow